+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All Notable changes to `League\Uri\Components` will be documented in this file
 - `Host::replace` now accepts negative offset like PHP 7.1+
 - `HierarchicalPath::without` argument is more strict accepts only array of integers
 - `Host::without` argument is more strict accepts only array of integers
+- `Query::merge` only accepts string
 
 ### Deprecated